+++ com.unity.render-pipelines.high-definition/HDRP/Material/Lit/Lit.cs
@@ -29,11 +29,12 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         };
 
         [GenerateHLSL]
-        public enum RefractionSSRayModel
+        public enum ProjectionModel
         {
             None = 0,
             Proxy = 1,
-            HiZ = 2
+            HiZ = 2,
+            Linear = 3
         };
 
         [GenerateHLSL]