+++ com.unity.render-pipelines.lightweight/LWRP/LightweightPipelineCore.cs
@@ -1,23 +1,11 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine.Rendering;
+using UnityEngine.Rendering.PostProcessing;
+using UnityEngine.XR;
 
 namespace UnityEngine.Experimental.Rendering.LightweightPipeline
 {
-    [Flags]
-    public enum FrameRenderingConfiguration
-    {
-        None                             = (0 << 0),
-        Stereo                           = (1 << 0),
-        Msaa                             = (1 << 1),
-        BeforeTransparentPostProcess     = (1 << 2),
-        PostProcess                      = (1 << 3),
-        DepthPrePass                     = (1 << 4),
-        DepthCopy                        = (1 << 5),
-        DefaultViewport                  = (1 << 6),
-        IntermediateTexture              = (1 << 7)
-    }
-
     [Flags]
     public enum PipelineCapabilities
     {
@@ -26,6 +14,68 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
         DirectionalShadows  = (1 << 2),
         LocalShadows        = (1 << 3),
         SoftShadows         = (1 << 4),
+    }
+    public enum MixedLightingSetup
+    {
+        None,
+        ShadowMask,
+        Subtractive,
+    };
+
+    public struct RenderingData
+    {
+        public CameraData cameraData;
+        public LightData lightData;
+        public ShadowData shadowData;
+        public bool supportsDynamicBatching;
+    }
+
+    public struct LightData
+    {
+        public int pixelAdditionalLightsCount;
+        public int totalAdditionalLightsCount;
+        public int mainLightIndex;
+        public List<VisibleLight> visibleLights;
+        public List<int> visibleLocalLightIndices;
+    }
+
+    public struct CameraData
+    {
+        public Camera camera;
+        public float renderScale;
+        public int msaaSamples;
+        public bool isSceneViewCamera;
+        public bool isDefaultViewport;
+        public bool isOffscreenRender;
+        public bool isHdrEnabled;
+        public bool requiresDepthTexture;
+        public bool requiresSoftParticles;
+        public bool requiresOpaqueTexture;
+        public Downsampling opaqueTextureDownsampling;
+
+        public bool isStereoEnabled;
+
+        public float maxShadowDistance;
+        public bool postProcessEnabled;
+        public PostProcessLayer postProcessLayer;
+    }
+
+    public struct ShadowData
+    {
+        public bool renderDirectionalShadows;
+        public bool requiresScreenSpaceShadowResolve;
+        public int directionalShadowAtlasWidth;
+        public int directionalShadowAtlasHeight;
+        public int directionalLightCascadeCount;
+        public Vector3 directionalLightCascades;
+        public bool renderLocalShadows;
+        public int localShadowAtlasWidth;
+        public int localShadowAtlasHeight;
+        public bool supportsSoftShadows;
+        public int bufferBitCount;
+
+        public LightShadows renderedDirectionalShadowQuality;
+        public LightShadows renderedLocalShadowQuality;
     }
 
     public class CameraComparer : IComparer<Camera>
@@ -36,7 +86,7 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
         }
     }
 
-    public class LightweightKeywords
+    public static class LightweightKeywords
     {
         public static readonly string AdditionalLightsText = "_ADDITIONAL_LIGHTS";
         public static readonly string VertexLightsText = "_VERTEX_LIGHTS";
@@ -46,6 +96,9 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
         public static readonly string LocalShadowsText = "_LOCAL_SHADOWS_ENABLED";
         public static readonly string SoftShadowsText = "_SHADOWS_SOFT";
         public static readonly string CascadeShadowsText = "_SHADOWS_CASCADE";
+        public static readonly string DepthNoMsaa = "_DEPTH_NO_MSAA";
+        public static readonly string DepthMsaa2 = "_DEPTH_MSAA_2";
+        public static readonly string DepthMsaa4 = "_DEPTH_MSAA_4";
 
 #if UNITY_2018_2_OR_NEWER
         public static readonly ShaderKeyword AdditionalLights = new ShaderKeyword(AdditionalLightsText);
@@ -108,39 +161,48 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
         {
             s_PipelineCapabilities = 0U;
 
-            if (pipelineAsset.MaxPixelLights > 1 || pipelineAsset.SupportsVertexLight)
-                s_PipelineCapabilities |= PipelineCapabilities.AdditionalLights;
+            // Strip variants based on selected pipeline features
+            if (!pipelineAsset.customShaderVariantStripping)
+            {
+                if (pipelineAsset.maxPixelLights > 1 || pipelineAsset.supportsVertexLight)
+                    s_PipelineCapabilities |= PipelineCapabilities.AdditionalLights;
 
-            if (pipelineAsset.SupportsVertexLight)
-                s_PipelineCapabilities |= PipelineCapabilities.VertexLights;
+                if (pipelineAsset.supportsVertexLight)
+                    s_PipelineCapabilities |= PipelineCapabilities.VertexLights;
 
-            if (pipelineAsset.SupportsDirectionalShadows)
-                s_PipelineCapabilities |= PipelineCapabilities.DirectionalShadows;
+                if (pipelineAsset.supportsDirectionalShadows)
+                    s_PipelineCapabilities |= PipelineCapabilities.DirectionalShadows;
 
-            if (pipelineAsset.SupportsLocalShadows)
-                s_PipelineCapabilities |= PipelineCapabilities.LocalShadows;
+                if (pipelineAsset.supportsLocalShadows)
+                    s_PipelineCapabilities |= PipelineCapabilities.LocalShadows;
 
-            bool anyShadows = pipelineAsset.SupportsDirectionalShadows || pipelineAsset.SupportsLocalShadows;
-            if (pipelineAsset.SupportsSoftShadows && anyShadows)
-                s_PipelineCapabilities |= PipelineCapabilities.SoftShadows;
+                bool anyShadows = pipelineAsset.supportsDirectionalShadows || pipelineAsset.supportsLocalShadows;
+                if (pipelineAsset.supportsSoftShadows && anyShadows)
+                    s_PipelineCapabilities |= PipelineCapabilities.SoftShadows;
+            }
+            else
+            {
+                if (pipelineAsset.keepAdditionalLightVariants)
+                    s_PipelineCapabilities |= PipelineCapabilities.AdditionalLights;
+
+                if (pipelineAsset.keepVertexLightVariants)
+                    s_PipelineCapabilities |= PipelineCapabilities.VertexLights;
+
+                if (pipelineAsset.keepDirectionalShadowVariants)
+                    s_PipelineCapabilities |= PipelineCapabilities.DirectionalShadows;
+
+                if (pipelineAsset.keepLocalShadowVariants)
+                    s_PipelineCapabilities |= PipelineCapabilities.LocalShadows;
+
+                if (pipelineAsset.keepSoftShadowVariants)
+                    s_PipelineCapabilities |= PipelineCapabilities.SoftShadows;
+            }
         }
 
         public static void DrawFullScreen(CommandBuffer commandBuffer, Material material,
             MaterialPropertyBlock properties = null, int shaderPassId = 0)
         {
             commandBuffer.DrawMesh(fullscreenMesh, Matrix4x4.identity, material, 0, shaderPassId, properties);
-        }
-
-        public static void StartStereoRendering(Camera camera, ref ScriptableRenderContext context, FrameRenderingConfiguration renderingConfiguration)
-        {
-            if (CoreUtils.HasFlag(renderingConfiguration, FrameRenderingConfiguration.Stereo))
-                context.StartMultiEye(camera);
-        }
-
-        public static void StopStereoRendering(Camera camera, ref ScriptableRenderContext context, FrameRenderingConfiguration renderingConfiguration)
-        {
-            if (CoreUtils.HasFlag(renderingConfiguration, FrameRenderingConfiguration.Stereo))
-                context.StopMultiEye(camera);
         }
 
         public static void GetLightCookieMatrix(VisibleLight light, out Matrix4x4 cookieMatrix)
@@ -183,6 +245,15 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
             // Remaining light types don't support cookies
         }
 
+        public static void CopyTexture(CommandBuffer cmd, RenderTargetIdentifier source, RenderTargetIdentifier dest, Material material)
+        {
+            // TODO: In order to issue a copyTexture we need to also check if source and dest have same size
+            //if (SystemInfo.copyTextureSupport != CopyTextureSupport.None)
+            //    cmd.CopyTexture(source, dest);
+            //else
+                cmd.Blit(source, dest, material);
+        }
+
         public static bool IsSupportedShadowType(LightType lightType)
         {
             return lightType == LightType.Directional || lightType == LightType.Spot;
@@ -191,15 +262,6 @@ namespace UnityEngine.Experimental.Rendering.LightweightPipeline
         public static bool IsSupportedCookieType(LightType lightType)
         {
             return lightType == LightType.Directional || lightType == LightType.Spot;
-        }
-
-        public static bool PlatformSupportsMSAABackBuffer()
-        {
-#if UNITY_ANDROID || UNITY_IPHONE || UNITY_TVOS || UNITY_SAMSUNGTV
-            return true;
-#else
-            return false;
-#endif
         }
     }
 }